+++ middleware/middleware_test.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/komuw/ong/log"
 
-	"github.com/akshayjshah/attest"
+	"go.akshayshah.org/attest"
 	"go.uber.org/goleak"
 )