+++ middleware/security_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/akshayjshah/attest"
+	"go.akshayshah.org/attest"
 )
 
 const nonceHeader = "CUSTOM-CSP-NONCE-TEST-HEADER"